refactor: pull some sortable table creation code out of clusternetwork.js
